adapt to query-params example
